==> R/01_load.qmd <==
---
title: "01_load"
author: "Group12"
format: html
editor: visual
---

## Clear Environment

```{r}
# | echo: true
# | eval: true

rm(list = ls())
```

## Load libraries

```{r}
# | echo: true
# | eval: true

# Loading relevant packages used in this section

library(tidyverse)
library(readxl)
library(readr)

```

## Load data

```{r}
# | echo: true
# | eval: true

# define path to to the excel file
path_to_xlsx_file <- "../data/_raw/PCOS_data_without_infertility.xlsx"

# list excel sheet names
sheet_names <- excel_sheets(path_to_xlsx_file)

# load file of interest by sheet name
PCOS_data_xlsx <- read_excel(path_to_xlsx_file,
                             sheet = "Full_new")

### quality check
print(head(PCOS_data_xlsx))

print(paste("Xlsx file saved at:",
            path_to_xlsx_file))

```

## Write and save to TSV

```{r}
# | echo: true
# | eval: true

# defining output path
path_to_tsv_file <- "../data/PCOS_data_without_infertility.tsv"

# read sheet of interest by sheet name
PCOS_data_tsv <- read_excel(path_to_xlsx_file,
                            sheet = "Full_new")

# save the data to TSV format
write.table(PCOS_data_tsv, file = path_to_tsv_file,
            sep = "\t",
            row.names = FALSE,
            quote = FALSE)

# quality check
print(paste("TSV file saved at:",
            path_to_tsv_file))
```
