==> R/02_clean.qmd <==
---
title: "02_clean"
author: "Group12"
format: html
editor: visual
---

## Clear environment

```{r}
# | echo: true
# | eval: true

rm(list = ls())

```

## Load libraries

```{r}
# | echo: true
# | eval: true

library(tidyverse)
```

## Load data

```{r}
# | echo: true
# | eval: true

# Load the file
PCOS <- read_tsv(file = "../data/PCOS_data_without_infertility.tsv")

```

## Data wrangling - Column removal of non relevant data columns

```{r}
# | echo: true
# | eval: true

#Removal of columns
PCOS <- PCOS |>
  select(-"Sl. No") |> 
  select(-"Marraige Status (Yrs)") |> 
  select(-"Fast food (Y/N)") |> 
  select(-"...45")

#QC
print(PCOS)
```

## Rename columns

```{r}
# | echo: true
# | eval: true

#renaming all columns
PCOS <- PCOS |> 
  rename("patient_ID" = "Patient File No.",
         "PCOS_diagnosis" = "PCOS (Y/N)",
         "age" = "Age (yrs)",
         "weight" = "Weight (Kg)",
         "height" = "Height(Cm)",
         "blood_type" = "Blood Group",
         "pulse" = "Pulse rate(bpm)",
         "breaths_pr_min" = "RR (breaths/min)",
         "hemaglobin" = "Hb(g/dl)",
         "cycle_RI" = "Cycle(R/I)",
         "period_length" = "Cycle length(days)",
         "pregnant" = "Pregnant(Y/N)",
         "abortions" = "No. of aborptions",
         "betaHCG_1" = "I   beta-HCG(mIU/mL)",
         "betaHCG_2" = "II    beta-HCG(mIU/mL)",
         "FSH" = "FSH(mIU/mL)",
         "LH" = "LH(mIU/mL)",
         "FSH_LH_ratio" = "FSH/LH",
         "hip" = "Hip(inch)",
         "waist" = "Waist(inch)",
         "waist_hip_ratio" = "Waist:Hip Ratio",
         "TSH" = "TSH (mIU/L)",
         "AMH" = "AMH(ng/mL)",
         "PRL" = "PRL(ng/mL)",
         "vitaminD3" = "Vit D3 (ng/mL)",
         "PRG" = "PRG(ng/mL)",
         "RBS" = "RBS(mg/dl)",
         "weight_gain" = "Weight gain(Y/N)",
         "hair_growth" = "hair growth(Y/N)",
         "darker_skin" = "Skin darkening (Y/N)",
         "hair_loss" = "Hair loss(Y/N)",
         "pimples" = "Pimples(Y/N)",
         "exercise" = "Reg.Exercise(Y/N)",
         "BP_systolic" = "BP _Systolic (mmHg)",
         "BP_diastolic" = "BP _Diastolic (mmHg)",
         "follicle_no_R" = "Follicle No. (L)",
         "follicle_no_L" = "Follicle No. (R)",
         "avg_fsize_L" = "Avg. F size (L) (mm)",
         "avg_fsize_R" = "Avg. F size (R) (mm)",
         "endometrium" = "Endometrium (mm)"
         ) 

#QC
print(PCOS)
```

## Correcting cells

Two cells were identified as inconsistent with the format - a misplaced period in betaHCG2 and a character in numeric column AMH. These will be changed to NA

```{r}
# | echo: true
# | eval: true

# Change values to NA
PCOS <- PCOS |> 
  mutate(betaHCG_2 = sub("\\.$", NA, betaHCG_2),
  AMH = ifelse(substr(AMH, 1, 1) %in% as.character(0:9), AMH, NA)) |>  
  mutate(betaHCG_2 = as.numeric(betaHCG_2),
         AMH = as.numeric(AMH))


# QC
print(PCOS)

# Checking for NA
PCOS |> 
  summarise(has_NA = any(is.na(AMH))) |> 
  pull(has_NA)


```

## Factorising column

```{r}
# | echo: true
# | eval: true


# Factorise columns
PCOS <- PCOS |> 
  mutate(pregnant = factor(pregnant, 
                           levels = c(0, 1), 
                           labels = c("No", "Yes")),
         weight_gain = factor(weight_gain, 
                              levels = c(0, 1), 
                              labels = c("No", "Yes")),
         hair_growth = factor(hair_growth, 
                              levels = c(0, 1), 
                              labels = c("No", "Yes")),
         hair_loss = factor(hair_loss, 
                            levels = c(0, 1), 
                            labels = c("No", "Yes")),
         darker_skin = factor(darker_skin, 
                              levels = c(0, 1), 
                              labels = c("No", "Yes")),
         pimples = factor(pimples, 
                          levels = c(0, 1), 
                          labels = c("No", "Yes")),
         exercise = factor(exercise, 
                           levels = c(0, 1), 
                           labels = c("No", "Yes")),
         PCOS_diagnosis = factor(PCOS_diagnosis,
                                 levels = c(0, 1),
                                 labels = c("No", "Yes")), 

         )

```

## Saving and Splitting the Data

```{r}
# | echo: true
# | eval: true

# Write the updated dataset into a new data file
write_tsv(x = PCOS,
          file = "../data/PCOS_clean.tsv")

# Check colnames of original dataset
colnames(PCOS)

## Read the data into separate files
# Load the dataset, selecting columns from defined ranges 
blood_measurements <- read_tsv(file = "../data/PCOS_clean.tsv",
                               col_select = c(1:2,
                                              10,
                                              15:19,
                                              23:28))

# View the column names of the selected data
colnames(blood_measurements)


# Select body measurements using column names (excluding blood measurements)
body_measurements <- read_tsv(file = "../data/PCOS_clean.tsv",
                              col_select = -c(10,
                                              15:19,
                                              23:28))

# View the column names of the selected data
colnames(body_measurements)


# Write the data into two data files
write_tsv(x = blood_measurements,
          file = "../data/PCOS_blood_measurements.tsv")
write_tsv(x = body_measurements,
          file = "../data/PCOS_body_measurements.tsv")

# Note: Patient No. and PCOS Y/N in both datasets!
```
