==> R/03_augment.qmd <==
---
title: "03_augment"
author: "Group12"
format: html
editor: visual
---

## Load libraries

```{r}
# | echo: true
# | eval: true

library(tidyverse)
library(dplyr)

```

## Load data

```{r}
# | echo: true
# | eval: true

# Load the file
body_measurements <- read_tsv(file = "../data/PCOS_body_measurements.tsv")
blood_measurements <- read_tsv(file = "../data/PCOS_blood_measurements.tsv")

```

## Unit changes

```{r}
# | echo: true
# | eval: true

# Going from inches to cm + new hip/waist ratio
body_measurements <- body_measurements |> 
  mutate(hip = hip * 2.54) |>
  mutate(waist = waist * 2.54) |>
  mutate(waist_hip_ratio = round(waist_hip_ratio * 2.54, 1))
  
```

## Rounding BMI and dividing into categories

```{r}
# | echo: true
# | eval: true

# Rounding of BMI and dividing into categories
body_measurements <- body_measurements |>
  mutate(BMI = round(BMI, 1)) |> 
  mutate(BMI_class = case_when(
    BMI < 18.5 ~ "Underweight",
    BMI <= 18.5 | BMI < 25 ~ "Normal weight",
    BMI <= 25 | BMI < 30 ~ "Overweight",
    BMI >= 30 ~ "Obesity")) |>
  mutate(BMI_class = factor(BMI_class,
                            levels =  c("Underweight", "Normal weight", 
                                        "Overweight", "Obesity"))) |>
  relocate(BMI_class, .after = BMI)
```

## Grouping ages

```{r}
# | echo: true
# | eval: true

# Grouping into different age groups
body_measurements <- body_measurements |>
  mutate(age_group = case_when(
    age < 20 ~ "]20",
    age >= 20 & age < 25 ~ "[20;25[",
    age >= 25 & age < 30 ~ "[25;30[",
    age >= 30 & age < 35 ~ "[30;35[",
    age >= 35 & age < 40 ~ "[35;40[",
    age >= 40 & age < 45 ~ "[40;45[",
    age >= 45 ~ "[45",)) |>
  mutate(age_group = factor(age_group,
                            levels =  c("]20", "[20;25[", "[25;30[", "[30;35[",
                                        "[35;40[", "[40;45[", "[45"))) |>
  relocate(age_group, .after = age)
```

## Change blood type

```{r}
# | echo: true
# | eval: true

# Change blood type into indicators
body_measurements <- body_measurements |>
  mutate(blood_type = case_when(
    blood_type == 11 ~ "A+",
    blood_type == 12 ~ "A-",
    blood_type == 13 ~ "B+",
    blood_type == 14 ~ "B-",
    blood_type == 15 ~ "O+",
    blood_type == 16 ~ "O-",
    blood_type == 17 ~ "AB+",
    blood_type == 18 ~ "AB-")) |> 
  mutate(blood_type = factor(blood_type, 
                             levels = c("A+", "A-", "B+", "B-", "O+", "O-",
                                        "AB+", "AB-")))

```

## Change cycle status into strings

```{r}
# | echo: true
# | eval: true

# Change cycle status from numbers to strings
body_measurements <- body_measurements |>
  mutate(cycle_RI = case_when(
    cycle_RI == 2 ~ "regular",
    cycle_RI == 4 ~ "irregular")) |> 
  mutate(cycle_RI = factor(cycle_RI, 
                           levels = c("irregular", "regular")))

```

## Creating a cycle/pregnancy stage column

```{r}
# | echo: true
# | eval: true

# create column with cycle/pregnancy stages  
blood_measurements <- blood_measurements |>
  mutate(cycle_or_pregnancy_stage = case_when(
    PRG <= 0.7 ~ "Follicular stage",   
    PRG > 0.7 & PRG <= 2 ~ "Likely follicular stage",                     
    PRG > 2 & PRG <= 10 ~ "Luteal stage",                    
    PRG > 10 & PRG <= 25 ~ "Luteal stage/possibly first or second trimester pregnancy",
    PRG > 25 & PRG <= 44 ~ "First or second trimester pregnancy", 
    PRG > 44 & PRG <= 65 ~ "Second trimester pregnancy",     
    PRG > 65 & PRG <= 82.5 ~ "Second or third trimester pregnancy",      
    PRG > 82.5 & PRG <= 290 ~ "Third trimester pregnancy",     
    TRUE ~ "Unknown" )) |> 
  mutate(cycle_or_pregnancy_stage = factor(cycle_or_pregnancy_stage, 
                                           levels= c("Follicular stage", 
                                                     "Likely follicular stage", 
                                                     "Luteal stage",
                                                     "Luteal stage/possibly first or second trimester pregnancy",
                                                     "First or second trimester pregnancy",
                                                     "Second trimester pregnancy",
                                                     "Second or third trimester pregnancy", 
                                                     "Third trimester pregnancy",
                                                     "Unknown")))

# QC
head(blood_measurements)
```

## Save to tsv files

```{r}
# | echo: true
# | eval: true

# write data into data files
write_tsv(x = blood_measurements,
          file = "../data/PCOS_blood_measurements_aug.tsv")
write_tsv(x = body_measurements,
          file = "../data/PCOS_body_measurements_aug.tsv")

# joining the dataframes (for educational purposes)
PCOS_merged <- inner_join(x = body_measurements,
                          y = blood_measurements,
                          by = join_by(patient_ID, PCOS_diagnosis))
write_tsv(x = PCOS_merged,
          file = "../data/PCOS_merged.tsv")

```
