==> R/04_describe.qmd <==
---
title: "04_describe"
author: "Group12"
format: html
editor: visual
---

## Clear environment

```{r}
# | echo: true
# | eval: true

rm(list = ls())
```

## Load libraries

```{r}
# | echo: true
# | eval: true

library(tidyverse)
library(RColorBrewer)
library(ggplot2)
library(ggridges)
library(table1)
```

## Load data

```{r}
# | echo: true
# | eval: true

#Load the file
body_measurements <- read_tsv(file = "../data/PCOS_body_measurements_aug.tsv")

blood_measurements <- read_tsv(file = "../data/PCOS_blood_measurements_aug.tsv")

PCOS <- read_tsv(file = "../data/PCOS_merged.tsv")

```

## Data statistics

Checking dimensions of metadata tibble.

```{r}
# | echo: true
# | eval: true

# Cheking the dimensions of the body measurements
body_measurements_dim <- body_measurements |>
  dim() |>
  tibble() |>
  rename("Body measurments dimensions" = "dim(body_measurements)") |>
  print()

# Cheking the dimensions of the blood measurements
blood_measurements_dim <- blood_measurements |>
  dim() |>
  tibble() |>
  rename("Blood measurments dimensions" = "dim(blood_measurements)") |>
  print()
  
# Cheking the dimensions of the blood measurements
PCOS_dim <- PCOS |>
  dim() |>
  tibble() |>
  rename("PCOS dimensions" = "dim(PCOS)") |>
  print()


```

Checking column names and count patients with and without PCOS diagnosis

```{r}
# | echo: true
# | eval: true

# which column names exist in the merged data set:
PCOS |> 
  colnames()

# counting patients with (yes) and without (no PCOS diagnosis)
PCOS |> 
  count(PCOS_diagnosis)


```

There are 364 patients without PCOS diagnosis and 177 who have been diagnosed with PCOS. There is an over representation of people without the diagnosis this will effects the results of the analysis, but this will not be taken into account.

### Count of distribution in cycle or pregnancy stages

```{r}
# | echo: true
# | eval: true 

blood_measurements |>    
  count(cycle_or_pregnancy_stage)
```

The cycle or pregnancy stage groups created earlier have a very skewed distribution between the groups.

## Plotting distributions of data

### Histogram of age distribution

```{r}
# | echo: true
# | eval: true

# Creating barplot of age distribution in the dataset
age_plot <- body_measurements |> 
  ggplot(mapping = aes(x = age, 
                       fill = PCOS_diagnosis)) +
  geom_histogram(position = position_dodge(preserve = "single"),
           alpha = 0.45,
           color = "black",
           binwidth = 5) +
  labs(x = "Age (binwidth = 5)", 
       y = "Count",
       title = "Histogram of age distribution stratified for PCOS diagnosis",
       fill = "PCOS diagnosis") +
  scale_fill_brewer(palette = "Set2") +
  theme_light() +
  theme(title = element_text(size = 14),
        legend.position = "bottom")


# Saving the plot as a picture
ggsave(age_plot,
       filename = "../results/04_age_hist.png",
       device = "png",
       height = 20,
       width = 20,
       units = "cm")
```

The age looks normally distributed

## Bar chart of BMI groups

```{r}
# | echo: true
# | eval: true

# Creating bar chart of BMI distribution stratified on PCOS diagnosis
BMI_bar <- body_measurements |>
  mutate(BMI_class = fct_relevel(BMI_class, "Underweight", "Normal weight", "Overweight",
                                 "Obesity")) |> 
  ggplot(mapping = aes(x = BMI_class,
                       fill = PCOS_diagnosis)) +
  geom_bar(position = position_dodge(preserve = "single"),
           alpha = 0.45,
           color = "black") +
  labs(x = "BMI groups", 
       y = "Count",
       title = "Bar chart over BMI groups stratified for PCOS diagnosis",
       fill = "PCOS diagnosis") +
  scale_fill_brewer(palette = "Set2") +
  theme_light() +
  theme(title = element_text(size = 14),
        legend.position = "bottom")

# Saving the plot as a picture
ggsave(BMI_bar,
       filename = "../results/04_BMI_bar.png",
       device = "png",
       height = 20,
       width = 20,
       units = "cm")

```

There is an over representation of data points in the Normal weight category and a scewed proportion of PCOS diagnoses in the normal weight category.

### Bar chart of blood groups

```{r}
# | echo: true
# | eval: true

# Creating barplot of blood types present in the dataset
blood_type_plot <- body_measurements |> 
  ggplot(mapping = aes(x = blood_type, 
                       fill = PCOS_diagnosis)) +
  geom_bar(position = position_dodge(preserve = "single"),
           alpha = 0.45,
           color = "black") +
  labs(x = "Blood type", 
       y = "Count",
       title = "Bar chart of blood groups stratified for PCOS diagnosis",
       fill = "PCOS diagnosis") +
  scale_fill_brewer(palette = "Set2") +
  theme_light() +
  theme(title = element_text(size = 14),
        legend.position = "bottom")



# Saving the plot as a picture
ggsave(blood_type_plot,
       filename = "../results/04_blood_type_bar.png",
       device = "png",
       height = 20,
       width = 20,
       units = "cm")

```

There is not an equal distribution of the data points between different blood groups.

### Density plot of period length

```{r}
# | echo: true
# | eval: true


period_plot <- body_measurements |> 
  ggplot(mapping = aes(x = period_length, 
                       fill = PCOS_diagnosis)) + 
  geom_density(alpha = 0.45) +
  labs(x = "Period length (days)", 
       y = "Density",
       title = "Density plot of period lengths",
       fill = "PCOS diagnosis") +
  scale_fill_brewer(palette = "Set2") +
  theme_light() +
  theme(title = element_text(size = 14),
        legend.position = "bottom")

# Saving the plot as a picture
ggsave(period_plot,
       filename = "../results/04_period_dens.png",
       device = "png",
       height = 20,
       width = 20,
       units = "cm")

```

The period length of people with PCOS diagnosis is more scattered then the lengths of people with no diagnosis.

### Summary table(s)

For easy summary of the data, the function table1 can be performed.

```{r}
# | echo: true
# | eval: true

# general summary table
summary_table <- PCOS |>
  table1(x = formula(~ age + weight + height + pulse + breaths_pr_min + cycle_RI +
                       period_length + period_length + pregnant + abortions + weight_gain +
                       hair_growth + darker_skin + hair_loss | PCOS_diagnosis)) 

# checking the summary table
print(summary_table)

# summary table for period related parameters
summary_period <- PCOS |>
  table1(x = formula(~ cycle_RI +  period_length + pregnant + abortions | PCOS_diagnosis)) 

# checking the summary table for period paramters
print(summary_period)

# summary table for symptoms often related to PCOS
summary_symptoms <- PCOS |>
  table1(x = formula(~ weight_gain + hair_growth + darker_skin + hair_loss | PCOS_diagnosis)) 

# checking the summary table for PCOS related symptoms
print(summary_symptoms)

```
