==> R/05_analysis_1.qmd <==
---
title: "05_analysis_1"
author: "Group12"
format: html
editor: visual
---

## Clear environment

```{r}
# | echo: true
# | eval: true

rm(list = ls())
```

## Load libraries

```{r}
# | echo: true
# | eval: true

# load libraries
library(tidyverse)
library(patchwork)
library(RColorBrewer)
library(ggplot2)
library(ggridges)
library(RColorBrewer)

```

## Load data

```{r}
# | echo: true
# | eval: true

# Loading file
body_measurements <- read_tsv(file = "../data/PCOS_body_measurements_aug.tsv")

```

# Analysis of body measurement data

First part of this project, will be to test if there is a correlation between patients with PCOS using the body measurements data.

## Follicle and PCOS

```{r}
# | echo: true
# | eval: true

# distribution of follicles in both ovaries
p_follicles_ovaries <- body_measurements |> 
  ggplot(mapping = aes(x = follicle_no_R,
                       y = follicle_no_L,
                       color = PCOS_diagnosis)) + 
  geom_point() +  
  geom_smooth(method = "lm", 
              se = TRUE, 
              aes(fill = PCOS_diagnosis), 
              alpha = 0.3) +
  labs(title = "Distribution of follicles in both ovaries",
       x = "Follicle No. (R)", 
       y = "Follicle No. (L)",
       color = "PCOS Diagnosis",
       fill = "PCOS Diagnosis") +
  scale_color_brewer(palette = "Set2") +
  scale_fill_brewer(palette = "Set2") +
  theme_light() +
  theme(title = element_text(size = 14),
        legend.position = "bottom")

# Saving the plot as a picture
ggsave(p_follicles_ovaries,
       filename = "../results/05_follicle_correlation.png",
       device = "png",
       height = 20,
       width = 20,
       units = "cm")

```

The number of follicles in the left and right ovaries are increased for people with the PCOS diagnosis. We look further into the number of follicles in each ovary:

```{r}
# | echo: true
# | eval: true

# distribution of follicles to the left
p_follicle_L <- body_measurements |> 
  ggplot(aes(x = PCOS_diagnosis, 
             y = follicle_no_L)) + 
  geom_boxplot(aes(fill = PCOS_diagnosis), 
               outlier.shape = NA) + 
  geom_jitter(width = 0.2, 
              alpha = 0.2, 
              size = 1) +
  labs(x = "PCOS (Y/N)", 
       y = "Follicle No. (L)",
       title = "Follicle number of left ovary and correlation to PCOS diagnosis",
       fill = "") + 
  scale_fill_brewer(palette = "Set2") +
  theme_light() +
  theme(title = element_text(size = 14),
        legend.position = "none")


# saving the plot
ggsave(p_follicle_L,
       filename = "../results/05_follicle_L_box.png",
       device = "png",
       height = 20,
       width = 20,
       units = "cm")

# distribution of follicles to the right
p_follicle_R <- body_measurements |> 
  ggplot(aes(x = PCOS_diagnosis, 
             y = follicle_no_R)) + 
  geom_boxplot(aes(fill = PCOS_diagnosis), 
               outlier.shape = NA) + 
  geom_jitter(width = 0.2, 
              alpha = 0.2, 
              size = 1) + 
  labs(x = "PCOS (Y/N)", 
       y = "Follicle No. (R)",
       title = "Follicle number of right ovary and correlation to PCOS diagnosis",
       fill = "") + 
  scale_fill_brewer(palette = "Set2") +
  theme_light() +
  theme(title = element_text(size = 14),
        legend.position = "none")


# saving the plot
ggsave(p_follicle_R,
       filename = "../results/05_follicle_R_box.png",
       device = "png",
       height = 20,
       width = 20,
       units = "cm")

```

In the two boxplots a significant increase in the number of follicles for both sides for people diagnosed with PCOS.

## Correlation between BMI and PCOS

```{r}
# | echo: true
# | eval: true

# boxplot of BMI and PCOS diagnosis
PCOS_BMI_correlation_box <- body_measurements |> 
  ggplot(mapping = aes(x = PCOS_diagnosis,
                       y = BMI,
                       fill = PCOS_diagnosis)) +
  geom_boxplot() +
  geom_jitter(width = 0.2, 
              alpha = 0.2, 
              size = 1) +
  labs(x = "PCOS diagnosis", 
       y = "BMI",
       title = "Correlation between BMI and PCOS diagnosis",
       fill = "PCOS diagnosis") +
  scale_fill_brewer(palette = "Set2") +
  theme_light() +
  theme(title = element_text(size = 14),
        legend.position = "bottom")

# Saving the plot as a picture
ggsave(PCOS_BMI_correlation_box,
       filename = "../results/05_PCOS_BMI_correlation_box.png",
       device = "png",
       height = 20,
       width = 20,
       units = "cm")


```

A slight difference can be seen between diagnosis and BMI.

## Correlation between PCOS and endometrium size

```{r}
# | echo: true
# | eval: true

# plotting the distribution of endometrium and PCOS diagnosis
PCOS_endometrium_correlation_box <- body_measurements |> 
  ggplot(mapping = aes(x = PCOS_diagnosis,
                       y = endometrium,
                       fill = PCOS_diagnosis)) +
  geom_boxplot() +
  geom_jitter(width = 0.2, 
              alpha = 0.2, 
              size = 1) +
  labs(x = "PCOS diagnosis", 
       y = "Endometrium size (mm)",
       title = "Correlation between endometrium size and PCOS diagnosis",
       fill = "PCOS diagnosis") +
  scale_fill_brewer(palette = "Set2") +
  theme_light() +
  theme(title = element_text(size = 14),
        legend.position = "bottom")


# Saving the plot as a picture
ggsave(PCOS_endometrium_correlation_box,
       filename = "../results/05_PCOS_endometrium_correlation_box.png",
       device = "png",
       height = 20,
       width = 20,
       units = "cm")

```

No difference can be seen in the endometrium size and PCIS diagnosis.

## Correlation between age and cycle stratified for PCOS

```{r}
# | echo: true
# | eval: true

# age vs. cycle stratified for PCOS diagnosis
age_cycle_plot <- na.omit(body_measurements) |> 
  ggplot(mapping = aes(x = age,
                       y = cycle_RI,
                       color = PCOS_diagnosis)) +
  geom_violin() +
  geom_jitter(width = 0.2, 
              alpha = 0.8, 
              size = 1) +
  labs(x = "Age", 
       y = "Cycle (Irregular / Regular)",
       title = "Correlation between age and cycle stratified for PCOS",
       color = "PCOS diagnosis") +
  scale_color_brewer(palette = "Set2") +
  theme_light() +
  theme(title = element_text(size = 14),
        legend.position = "bottom")


# Saving the plot as a picture
ggsave(age_cycle_plot,
       filename = "../results/05_age_cycle_plot.png",
       device = "png",
       height = 20,
       width = 20,
       units = "cm")

```

Unable to conclude if there is a similarity between people with PCOS, and what their ages is and if their cycle is regular/irregular.

## Age groups vs. hair growth stratified for PCOS

```{r}
# | echo: true
# | eval: true

p_age_hair <- body_measurements |> 
  ggplot(aes(x=hair_growth,
             fill = PCOS_diagnosis))+
  geom_bar(position = position_dodge2(preserve = "single"), width = 0.8,
           alpha = 0.8)+
  facet_wrap(~ age_group)+
  labs(x = "Hair growth (yes/no)", 
       y = "Count",
       title = "Bar plot of hair growth in different age groups stratisfied for PCOS diagnosis",
       fill = "PCOS diagnosis") +
  scale_fill_brewer(palette = "Set2") +
  theme_light() + 
  theme(title = element_text(size = 14),
        legend.position = "bottom",
        axis.title.x= element_text(size = 12))
  

# Saving the plot as a picture
ggsave(p_age_hair,
       filename = "../results/05_age_hair_correlation_plot.png",
       device = "png",
       height = 20,
       width = 20,
       units = "cm")
```

An increase in hair growth can not be seen in people with PCOS diagnosis. People with no diagnosis see no increase in hair growth.

## Age groups vs. pimples stratified for PCOS

```{r}
# | echo: true
# | eval: true

#bar plot 
p_age_pimple <- body_measurements |> 
  ggplot(aes(x=pimples,
             fill = PCOS_diagnosis))+
  geom_bar(position = position_dodge2(preserve = "single"), width = 0.8,
           alpha = 0.8)+
  facet_wrap(~ age_group)+
  labs(x = "Pimples (yes/no)", 
       y = "Count",
       title = "Bar plot of pimples in different age groups stratisfied for PCOS diagnosis",
       fill = "PCOS diagnosis") +
  scale_fill_brewer(palette = "Set2") +
  theme_light() + 
  theme(title = element_text(size = 14),
        legend.position = "bottom",
        axis.title.x= element_text(size = 10))


# Saving the plot as a picture
ggsave(p_age_pimple,
       filename = "../results/05_age_pimple_correlation_plot.png",
       device = "png",
       height = 20,
       width = 20,
       units = "cm")
```

People with PCOS diagnosis experience more pimple then those with no PCOS diagnosis.
