==> R/08_logistic_regression.qmd <==
---
title: "08_logistic_regression.qmd"
author: "Group12"
format: html
editor: visual
---

## Logistic regression

## Clear environment

```{r}
# | echo: true 
# | eval: true  

rm(list = ls())
```

## Load libraries

```{r}
# | echo: true 
# | eval: true  

library(tidyverse)
library(ggfortify)

```

## Load data

```{r}
# | echo: true 
# | eval: true 

body_measurements <- read_tsv(file = "../data/PCOS_body_measurements_aug.tsv")
blood_measurements <- read_tsv(file = "../data/PCOS_blood_measurements_aug.tsv")

# verifying column names
colnames(blood_measurements)
```

## Logistic regression

From previous analysis, the follicle number and size seems slight important for individuals with PCOS diagnosis compared to individuals without. Out of curiosity we wanted to see if we could perform a linear regression model on a subset of the data

```{r}
# | echo: true 
# | eval: true 

# wrangling data for analysis with logistic regression
body_measurements <- body_measurements |>
  mutate(PCOS_diagnosis = factor(PCOS_diagnosis,
                                 levels = c("No", "Yes"),
                                 labels = c(0, 1)))|>
  mutate(PCOS_diagnosis = as.numeric(PCOS_diagnosis) -1)

# making a new dataframe
data_model <- body_measurements |>
  dplyr::select(PCOS_diagnosis, follicle_no_R, follicle_no_L, avg_fsize_R, avg_fsize_L)

# quality check
print(data_model)

# performing the model
follicle_model <- glm(PCOS_diagnosis ~., 
          family = "binomial",
          data = data_model)

summary(follicle_model)

```

Now we see that for number of follicles to the right and left, both have a low p-value, showing significant importance of the individuals with PCOS. However the size does not seem very important.

We will know try to predict on the binary data for follicle no to the right and left respectively.

```{r}
# | echo: true 
# | eval: true 

### follicle no to R

plot_follicle_R <- body_measurements |>
  ggplot(mapping = aes(follicle_no_R, PCOS_diagnosis)) + 
  geom_point(alpha = 0.45,
             color = "#1B9E77") + 
  geom_smooth(method = "glm",
              method.args = list(family = "binomial"),
              color = "#7570B3",
              alpha = 0.45) +
  geom_hline(yintercept = 0.5,
             linetype = "dashed",
             color = "#D95F02",
             linewidth = 0.5,
             alpha = 0.3) +  
  geom_vline(xintercept = 8.4,
             linetype = "dashed",
             color = "#D95F02",
             linewidth = 0.5,
             alpha = 0.3) + 
  geom_point(x = 8.4,
             y = 0.5,
             color = "#D95F02",
             size = 4,
             alpha = 0.3) + 
  labs(title = "Logistic Regression Model",
        x = "Follicle number (R)",
        y = "Probability of PCOS") +
   theme_light() +
  theme(title = element_text(size = 14))


print(plot_follicle_R)

# Saving the plot as a picture
ggsave(plot_follicle_R,
       filename = "../results/08_plot_follicle_R.png",
       device = "png",
       height = 20,
       width = 20,
       units = "cm")

### follicle no to L

plot_follicle_L <- body_measurements |>
  ggplot(mapping = aes(follicle_no_L, PCOS_diagnosis)) + 
  geom_point(alpha = 0.45,
             color = "#1B9E77") + 
  geom_smooth(method = "glm",
              method.args = list(family = "binomial"),
              color = "#7570B3",
              alpha = 0.45) +
  geom_hline(yintercept = 0.5,
             linetype = "dashed",
             color = "#D95F02",
             linewidth = 0.5,
             alpha = 0.3) +  
  geom_vline(xintercept = 9.0,
             linetype = "dashed",
             color = "#D95F02",
             linewidth = 0.5,
             alpha = 0.3) + 
  geom_point(x = 9.0,
             y = 0.5,
             color = "#D95F02",
             size = 4,
             alpha = 0.3) + 
  labs(title = "Logistic Regression Model",
        x = "Follicle number (L)",
        y = "Probability of PCOS") +
  theme_light() +
  theme(title = element_text(size = 14))


print(plot_follicle_L)

# Saving the plot as a picture
ggsave(plot_follicle_L,
       filename = "../results/08_plot_follicle_L.png",
       device = "png",
       height = 20,
       width = 20,
       units = "cm")


```

Now we want to test, if the difference between number of follicles in right and left, has any significance.

```{r}
# | echo: true 
# | eval: true 

body_measurements <- body_measurements |>
  mutate(follicle_no_diff = abs(follicle_no_R - follicle_no_L))

lm_diff <- glm(as.factor(PCOS_diagnosis) ~ follicle_no_diff, 
               family = "binomial",
               data = body_measurements)

summary(lm_diff)

plot_follicle_diff <- body_measurements |>
  ggplot(mapping = aes(follicle_no_diff, PCOS_diagnosis)) + 
  geom_point(alpha = 0.45,
             color = "#1B9E77") + 
  geom_smooth(method = "glm",
              method.args = list(family = "binomial"),
              color = "#7570B3",
              alpha = 0.45) +
  geom_hline(yintercept = 0.5,
             linetype = "dashed",
             color = "#D95F02",
             linewidth = 0.5,
             alpha = 0.3) +  
  geom_vline(xintercept = 3.8,
             linetype = "dashed",
             color = "#D95F02",
             linewidth = 0.5,
             alpha = 0.3) + 
  geom_point(x = 3.8,
             y = 0.5,
             color = "#D95F02",
             size = 4,
             alpha = 0.3)
   theme_light()

print(plot_follicle_diff)

# checking summary statistics
diff_summary <- body_measurements |>
   table1(x = formula(~ follicle_no_diff | PCOS_diagnosis)) 

print(diff_summary)

```

There is a significant difference between the number of follicles to the left and right.
